Move checking of trait / interface requirements to typing phase
Summary:Previously, we would check trait / interface requirements at declaration time,
and if there existed two requirements that were for the same class / interface,
we would call the subtyping function to pick the more specific one. However,
the subtype function needs the declaration information to be present, and there
is no guarantee that this is true while declaration is still ongoing.  This led
to a number of bugs. It also makes it hard to implement lazy decl, since if
these circular dependencies went undetected, they could crash the typechecker.

The bugs with the current implementation manifest both in missing and spurious
type errors. A missing type error can occur when a requirement references a
class that has not been declared yet:

  interface I1 { require extends Super<NotDeclaredYet>; }
  interface I2 { require extends Super<Foo>; }
  interface I3 extends I1, I2 {}

Since we were trying to pick the more specific Super<> type, we would check if
Super<Foo> was a subtype of Super<NotDeclaredYet>. But since NotDeclaredYet
cannot be retrieved, it defaults to the `any` type, and Super<Foo> is indeed a
subtype of Super<any>, so we declare `I3` as `require extends Super<Foo>`. The
unit test iface_require_circular.php demonstrates this -- prior to this diff,
it printed 'No errors'.

We would also see spurious type errors:

  class Super<+T> {}
  interface IA {}
  interface IB {}
  interface IC extends IA, IB {}
  interface I1 { require extends Super<IA>; }
  interface I2 { require extends Super<IB>; }
  interface I3 extends I1, I2 {}

`I3` can be successfully implemented by a class that extends `Super<IC>`, but
because we can only declare `I3` to `require extends` one of `Super<IA>` or
`Super<IB>`, the typechecker would raise an error because neither one is a
subtype of the other. `iface_require_covariant.php` demonstrates this.

Because of these issues, this diff does away with the logic that tries to pick
a more specific requirement. Only when the interface gets implemented do we
check for consistency. This means that some unimplementable interfaces no
longer raise errors:

  class Super<T> {} // note that T is now invariant
  interface I1 { require extends Super<int>; }
  interface I2 { require extends Super<float>; }
  interface I3 extends I1, I2 {}

But this doesn't really matter since the instant we try to implement `I3` the
typechecker would still point out the error, albeit in a way that can be
slightly harder to debug.

Performance
===========

Not being able to use unification & subtyping to narrow down the set of
requirements at declaration time does come with a slight performance penalty.
My initial implementation showed a ~4% regression in type-check time. However,
I realized that I could still narrow down the list of requirements without real
unification -- just checking that the requirements were syntactically equal was
enough in the majority of cases. With performance hack applied, the overhead is
now ~1%, which seems more acceptable.

Reviewed By: alexchow

Differential Revision: D2926316

fb-gh-sync-id: 0760c136b327aaf782e57511bcc31fb382164c18
shipit-source-id: 0760c136b327aaf782e57511bcc31fb382164c18