NastInitCheck should use get_class_dep
Summary:Currently it uses get_class, which only adds a recheck (but not a redecl)
dependency on the parent classes it accesses. This is clearly wrong because if
a parent class gains another property, the child classes will all have to rerun
decl to make sure they have initialized all the properties that they now
inherit.

In practice it doesn't really matter because we add decl dependencies from a
class to all its ancestors anyway in other parts of the code, but this is the
more correct thing to do.

Found it while doing my decl refactor.

Reviewed By: dlreeves

Differential Revision: D2968688

fb-gh-sync-id: 5125a2e6248c081ff7daa5a698c9461fb54075b3
shipit-source-id: 5125a2e6248c081ff7daa5a698c9461fb54075b3