Solve optimization issues on PPC64 by improving robustness
It's not a good idea to simply shut down optimizations if your code is
breaking because the compiler might change in the future. It's wiser
to make your code more robust in order to avoid falling in such cases
that optimization could break your code.

The fixup issue was already solved on upstream
(fa13910bd1f391f6110b1c07cd5a55d265e1b6be) but as discussed above, I
changed it to avoid the optimization issue by using a parent frame
(which tail call optimization doesn't eliminate).

The stub issue is due to register size visibility. On X64 it's possible
to look only at the lowest byte of a register directly when a bool value
is an argument. On PPC64 I faced 2 different behaviors:

  Without optimization: that volatile argument register was saved (1
byte) onto the stack temporarily. Next, loaded (1 byte) and by doing
that, it'd clear any undesired bits that the argument had. After that
the whole register is 32bits-compared in order to evaluate the boolean.
This was OK and we didn't run into troubles.

  With optimization: the volatile argument register is directly
evaluated as a boolean and 32bits-compared (note: more than 1 byte is
used). If the caller was a C++ function, then the compiler would take
care of all the 32bits to be consistent, however on our case, the JITted
code needs to implement that. On X64 this change now generates a "mov
$0x1,%ecx" instead of "mov $0x1,%cl", which I believe it's not a major
drawback.

Thanks to @tuliom for discussing these decisions.