Give ElmKey and PropAccessInfo external linkage
Summary:One of the restrictions of the type-scanner is that anything with a
custom scanner function defined must have external linkage. This also applies to
anything used as a template parameter to a template with a custom scanner
function (like the req-containers). This only (currently) affects two types:
ElmKey and PropAccessInfo. Give these types external linkage by defining them
within MixedArray and ObjectData, respectively.

Reviewed By: edwinsmith

Differential Revision: D2994611

fb-gh-sync-id: 1dccb2f911c2ff9f6990c3f9ebd6c1bc6d3631d2
shipit-source-id: 1dccb2f911c2ff9f6990c3f9ebd6c1bc6d3631d2