Unittest fixes when running with icu-53.1
Summary:Changes to 1810.php - The "UK_UA_ODESSA" locale is named "uk" as of ICU versions
53.1 and 56.1.  The "uk-ua_CALIFORNIA@currency=;currency=GRN" locale is not
valid as of 56.1, so it was just removed.

Changes to ucsdet.php - There was a test that asserted EncodingDetector would
find a string invalid, but ended up being valid utf16-be under icu-53.1. Change the string
to something both icu versions think are invalid. EncodingDetector doesn't seem to exist
in php7, so I didn't check it against icu-56.1

Reviewed By: markw65

Differential Revision: D2997664

fb-gh-sync-id: efe325c0b8cd5a8de052ce43f95d8b3289b0b10a
shipit-source-id: efe325c0b8cd5a8de052ce43f95d8b3289b0b10a