Correctly handle type_structure() resolution for type access
Summary:When resolving a `type_structure()` for a type constant involving a nullable
type access we were not correctly marking the resulting array as nullable. For
instance:

```
class Foo {
  const type X = int;
  const type Y = ?Foo::X;
}
```

The type structure for `Foo::Y` would be the same as `Foo::X` instead of being
marked as nullable.

The resolution code should probably be restructured to prevent this kind of bug
from occurring, but I believe this is the last instance of this bug and we
don't plan on adding any additional types so I'll go with the simple fix in
this case.

Reviewed By: paulbiss

Differential Revision: D3002951

fb-gh-sync-id: e920aae6c7d257c29455173212e4591a49bb5518
shipit-source-id: e920aae6c7d257c29455173212e4591a49bb5518