Use lists instead of sets to pass file names around in search service
Summary:Search service is another place that has a lot of global, but not shared (in the sense of "not in shared memory") state that will have to be synchronized from master typechecker process. It involves things like "set of all file names", which at our scale are pretty slow[1]. Lists seem to perform much better.

In the process of eliminating them, it turned out that they only exist because of merge operation in HackSearchService.update_search_index, which merges "files" with "php_files" to enable searching in non-Hack files too. But ever since Hack build "files" is a superset of "php_files" (there even is a comment describing it), so the merge operation is no-op, and the only place that uses php_files can be removed, along with it's generation in parser code.

Reviewed By: dlreeves, int3

Differential Revision: D2981059

fb-gh-sync-id: f3bdd673fb44cb3a91ea7a4dde74f456f4aa4674
shipit-source-id: f3bdd673fb44cb3a91ea7a4dde74f456f4aa4674