Replace hand-rolled vector implementations
Summary:In a few places we allocate a raw array of memory (without running any
constructors), tracking the number of constructed elements, and running placement
new and destructors manually as elements are added or removed from the
array. This is exactly what req::vector does, so just use that instead.

Reviewed By: markw65, edwinsmith

Differential Revision: D3006162

fb-gh-sync-id: 96ba944ba5b510a1aa7d2a595dff1d8a5d925d70
shipit-source-id: 96ba944ba5b510a1aa7d2a595dff1d8a5d925d70