Rewrite handleSRHelper as a unique stub
Summary:Kill the asm stub and replace it with a vasm stub.  Omits the CFI info,
since (a) MCGenerator::handleServiceRequest() is noexcept, so we don't
need it for unwinding; (b) we unwind through all TC frames at once with
our custom handler, so we don't need it for unwinding; and (c) it's
incorrect anyway, now that we spill to the native stack.

Reviewed By: swtaarrs

Differential Revision: D2965449

fb-gh-sync-id: b50f5256c5a30ade539356e5b7a8294cf6a4696b
shipit-source-id: b50f5256c5a30ade539356e5b7a8294cf6a4696b