Error list served from IDE process via persistent connection
Summary: Whenever we have new errors, or there are too many errors to efficiently check if they are new, send them to IDE process. The IDE process can then answer status calls immediately. In the future it will allow also for a subscription model, when we notify editors about new errors when they are ready.

Reviewed By: int3

Differential Revision: D2995775

fb-gh-sync-id: 337520440a6c1b9d54c5c3ebdb78ddef6b31b377
shipit-source-id: 337520440a6c1b9d54c5c3ebdb78ddef6b31b377