Expose typing coverage in persistent connection mode
Summary:I though that after find-refs I am done with calls that are currently used, but Nuclide started asking about coverage this week (#10215644).

peterhal: we should put some freeze on the new calls issued by Nuclide if we want to start making the switch to persistent connection and testing in the wild.

Reviewed By: alexchow

Differential Revision: D3007237

fb-gh-sync-id: d70860b04621b1314db1edb8456d1009dc8ac0b1
shipit-source-id: d70860b04621b1314db1edb8456d1009dc8ac0b1