Document vasm-instr.h
Summary:Vasm was designed to mirror x64, but now that we live in a multi-
targeting world, we should clarify which x64 instructions are for
general use, and which are for x64 lowering.

The instructions also lack documentation for the most part, so add some
(ignoring the floating-point instructions that I don't recognize... we
can document those in a separate diff, later).

Reviewed By: swtaarrs

Differential Revision: D3010901

fb-gh-sync-id: e5fe81bdadd3a8d21f706349c7a776ccf82e9ddf
shipit-source-id: e5fe81bdadd3a8d21f706349c7a776ccf82e9ddf