Fix unwinder/fcallArray-stub bug
Summary:The stub for CallArray uses a C++ helper to setup the ActRec for the call. The
helper will sometimes decide to abort the call and return null. In order to
return to the original caller the stub would therefore leave the return address
on the native stack.

The C++ helper also invokes function entry hook on success, which may in turn
raise an exception. This is problematic for two reasons: throwing from a stub
can leave the native stack unbalanced, and the catch trace for the CallArray
would be skipped preventing it from doing any necessary cleanup and possibly
leaving the native stack further unbalanced.

This diff more eagerly writes the return address into the newly created ActRec
and properly restores the native stack via a special case for the fcallArray
stub in the unwinder which uses a second stub to pop the extra 16 bytes from the
stack and restore the vmsp and vmfp from rds as they may have been updated in
the C++ helper.

Reviewed By: mxw

Differential Revision: D3011783

fb-gh-sync-id: e2496d0475c74cd91f560f44087c302fba0f6ac7
shipit-source-id: e2496d0475c74cd91f560f44087c302fba0f6ac7