Fixed divergence between hh_single_type_check and hhi files
Summary:There is a difference between the behaviour of hh_single_type_check and hh_client
on some of the builtin types (ImmMap, ConstMap, KeyedIterable, etc) that have variant
type parameters. Definitive behaviour is that of hh_client, which gets its definitions
from hhi files under hphp/hack/hhi. The fix is to update the literal "builtins" string
in hh_single_type_check.ml to make the variance annotations consistent with those in
the hhi files.

Reviewed By: int3

Differential Revision: D3018185

fb-gh-sync-id: a49a8f9f2fc5c5d92bf538a1d7ae36c71a2bbeac
shipit-source-id: a49a8f9f2fc5c5d92bf538a1d7ae36c71a2bbeac