Fix idx notices with strings
Summary:isset($string, $idx) reports true if $idx is bool or double, and
isset($string, (int)$idx) would have returned true. But if $idx is
bool or double, $string[$idx] issues a notice, and idx is not supposed
to do that.

Suppress the notice with an explicit cast.

I first tried only accepting integer indexes (and returning the
default for anything else), but that changes the behavior for
strictly-integer-strings, which are supposed to behave like the
corresponding integer. I could explicitly reject bools and doubles,
but this seemed simpler (and appart from suppressing the notice,
doesn't change the behavior).

Reviewed By: paulbiss

Differential Revision: D3019051

fb-gh-sync-id: 224880de80c7d46bb836e716167db7a303f3b07c
shipit-source-id: 224880de80c7d46bb836e716167db7a303f3b07c