Allow type hints for variadic function in Hack
Summary:When variadic parameters were added to HHVM in {D1235901} we initially banned declaring them with a type hint. This is because PHP specifies that the type hint must be enforced for each argument passed to the variadic method. HHVM wanted to avoid performing the O(n) check at runtime.

Now 2 years have passed and there is still no plans to support enforcing these type hints at runtime. Even if they are not enforced, the type hints are still valuable for Hack code because the type checker and use the information for its analysis. With that in mind I am proposing we lift the restriction on variadic parameters with type hints for Hack files, while maintaining the current restriction for PHP code.

The only change here is checking if a file enable HH syntax, and if so translate the type hint from something like `int ...` to `array<int>`.

Reviewed By: fredemmott

Differential Revision: D3004057

fb-gh-sync-id: 5745f6088f0bebdacdbd6f2390c8247d9ef18e33
shipit-source-id: 5745f6088f0bebdacdbd6f2390c8247d9ef18e33