Factor out decl mode code into a separate folder that is a dependency of typing:typing
Summary:This makes it possible for Typing_env to invoke Typing_decl without creating a
circular module dependency.

This disentangles a big part of our code, but it's still not a great situation.
The phases roughly go as follows:

1. Parsing (code in parsing/)
2. Global naming (code in naming/)
3. Type-decl (code in decl/ with this diff), which does decl-level naming (code in naming/), which produces records defined in Typing_defs (code in typing/)
4. Typechecking (code in typing/), which reads the Typing_defs records, and also does body-level naming (code in naming/)

Potential future plans include factoring out the Typing_defs into decl/ or
their own folder, and / or splitting naming/ into the decl and typing folders.
But neither of these block lazy-decl, so we can punt on them for now.

Reviewed By: alexchow

Differential Revision: D3008696

fb-gh-sync-id: 36383f58c8423b639f6080d9534edbb3dfdfba4b
shipit-source-id: 36383f58c8423b639f6080d9534edbb3dfdfba4b