Re-use 3.12 embed_systemlibs for hphpize'd extensions
This is necessary, as for hphpize'd extensions, there is no access to
HPHP_HOME. The extension also shouldn't need the repo-schema and compiler-id
embedded. Without this change, you can't "link" hphpize'd extensions, as they
fail with:

	[  1%] Linking CXX shared library mongodb.so
	Generating Repo Schema ID and Compiler ID
	/bin/sh: 1: cd: can't cd to /hphp/util
	CMakeFiles/mongodb.dir/build.make:2802: recipe for target 'mongodb.so' failed
	make[2]: *** [mongodb.so] Error 2
	CMakeFiles/Makefile2:99: recipe for target 'CMakeFiles/mongodb.dir/all' failed
	make[1]: *** [CMakeFiles/mongodb.dir/all] Error 2
	Makefile:116: recipe for target 'all' failed
	make: *** [all] Error 2

I would like not to have to have this definition copied, but I see no other way.

Note, that this does not work on Windows, but then again, hphpize'd extensions
did not work in 3.12 on Windows either, as it already referred to this unset
HPHP_HOME anyway.