Lazy decl without the laziness
Summary:This diff adds Typing_lazy_heap, which provides the lazy decl fallback behavior --
if a class is not found in the heap, we will go declare it if we know it
exists (by looking up its name in the naming heap).

Originally, this functionality was supposed to be part of
`Typing_env.get_class` and friends. However, I found a bunch of places in the
type-check code that were accessing the Typing_heap without going through
Typing_env. Initially, I thought that those were unintentional bugs, since it
would mean that the existing code would be accessing a class without recording
a dependency. In practice this is typically not an issue because some other
part of the typing code will call `get_class` and ensure that the dependency is
recorded. However, after I "fixed" the callsites to use `get_class`, I observed
significant slowdowns in fresh init. It turns out that adding dependencies --
at least when saved state is not present -- is really
expensive, and many of those direct-to-heap callsites were probably
intentional.

Saved state seems to mitigate most of this, but we'd still pay the cost during
rebases and during our development of hh_server, so this additional delay is not really
acceptable. Thus Typing_lazy_heap was born. (Note that this doesn't mean we've
wasted work in breaking the circular dependency of decl code on Typing_env;
Typing_env depends on Typing_lazy_heap which depends on decl code, so there
would still have been a circularity.)

However, we still run the full type-decl phase during init, so we should never
have a case where a class is missing in the typing heap but exists in the
naming heap. A few more steps are required before that can happen.

Here are the remainder (that I know of):
* Hack build needs to use the lazy heap
* Error handling. As of this diff, if we encounter a type error while doing
  decl, we will lump its filename into failed_check instead of failed_decl like
  we are supposed to. I think the solution would be to not put the decl data
  into the heap if we encounter type errors.

What this diff does achieve: you can now comment out the line in
`ServerInit.init` that invokes type-decl and start up hh_server without seeing
spurious type errors.

Reviewed By: dlreeves

Differential Revision: D3023639

fb-gh-sync-id: e9f3132183b49a01bb0c1c986d1e1579e2c957d9
shipit-source-id: e9f3132183b49a01bb0c1c986d1e1579e2c957d9