Add a vasm width-checking pass
Summary:This lets us make a stronger claim about our vasm instructions---namely,
that for a Vreg{8,16,32} def, what happens to the upper bits never
matters, and are irrelevant to the semantics of the instruction.  Before
this pass (and the changes it required), we could only say that those
upper bits never needed to be preserved, but /did/ occasionally have
required side-effects (e.g., movl{} needed to zero the upper bits, but
no longer does).

Since the LLVM backend is gone, we won't get these checks for free
anymore.  (LLVM also missed some spots, when it encountered instructions
in the unit that it didn't support.)

Reviewed By: swtaarrs

Differential Revision: D3005230

fb-gh-sync-id: f34379f0805bc11d92cb8a4568586901c49f698e
shipit-source-id: f34379f0805bc11d92cb8a4568586901c49f698e