Check for zero-length string in soap's encode_add_ns()
Summary:PHP will pass a const char* to this function, and if there's no
namespace, then it will have stored that string as nullptr.  We store the string
as a std::string, and when we pass in the result of std::string::c_str(), it'll
give us a non-nullptr empty string.

Found when using libxml2-2.9.1.

Reviewed By: swtaarrs

Differential Revision: D3037041

fb-gh-sync-id: a4ad2af732930045a553da098a7315440d72371d
shipit-source-id: a4ad2af732930045a553da098a7315440d72371d