Clarify stack offset operators and interconversions
Summary:Some of the stack offset operators don't make any sense to me, e.g.,

  - operator-(BCSPOffset).  The top of the interpreter eval stack is
    the end-all be-all for our eval stack.  A negative BCSPOffset
    doesn't seem meaningful.

  - operator-(int32_t, IRSPOffset).  How do you subtract out a relative
    value from an integer?

  - operator-(XOffset, YOffset).  Subtracting offsets that aren't
    relative to the same thing just doesn't make sense semantically.

Some of these operators are totally unused, so we delete them to avoid
future confusion.  The wacky difference operators, however, do serve a
purpose...

In particular, we occasionally want to take the offset of a stack slot
represented one way (say, an FPInvOffset) and convert it to another
representation (say, an IRSPOffset) relative to a stack pointer (which
is typically typed as an FPInvOffset)---or vice versa.

These relationships are pretty fundamental to these offset types
(they're even diagrammed above, in a nice ASCII diagram), and these are
the actual primitives we want where before we had cryptic subtraction.
So, add some member functions for them.  In practice, this seems to
make callsites clearer.

To that end, also kill toIRSPOffset().  This converts a BCSPOffset to
an FPInvOffset (given the "anchor" bytecode stack pointer offset) and
then to an IRSPOffset (given the anchor IR SP offset)---but its
argument order seems a bit fragile.  Doing the conversions explicitly
makes it more obvious what transformation is happening.

Reviewed By: swtaarrs

Differential Revision: D3035080

fb-gh-sync-id: ae44ce7f8fd4e5914227064e564da5c642472272
shipit-source-id: ae44ce7f8fd4e5914227064e564da5c642472272