Fix off-by-one in FrameStateMgr::stackState()
Summary:We're 1-indexing the FrameState::stack vector instead of 0-indexing
it.  Fortunately, we only ever index the vector via a single accessor
(FrameStateMgr::stackState()), which resized the vector as necessary.
Still, this meant our assertions were loose, and it was hiding real
bugs, and also making our stack offset accounting confusing, so let's
fix it.

Reviewed By: swtaarrs

Differential Revision: D3032714

fb-gh-sync-id: 53521f66f6d84ad6036c3ba34a8f016bbb086ff2
shipit-source-id: 53521f66f6d84ad6036c3ba34a8f016bbb086ff2