Port find-lvar-refs from hh_ide.js to IDE process
Summary: This is a function that exists only in the javascript version, with no server counterpart. See the linked task fro the plans/reasons for getting rid of javascript. The reply format is copypasta from hh_ide.js, with the intention to eventually change it, when we will merge find-lvar-refs/find-refs/find-class-refs into a single FindRefs call.

Reviewed By: int3

Differential Revision: D3048250

fb-gh-sync-id: 1c93c9787150c828737d877f6fa0e363e030032a
shipit-source-id: 1c93c9787150c828737d877f6fa0e363e030032a