Answer "hh_client --search" queries from IDE process, remove search index from typechecker process
Summary:The search index is duplicated in IDE and typechecker process, and it's a bigger memory waster than I expected. As a band aid, this diff adds a way to answer hh_client search query from the IDE process. It does it in an ugly way, with few "if SEARCH then ..." statements.

We've talked with Jez about porting all the commands to IDE process, but this is more complicated and out of scope for now. Main problems I foresee are:

- IDE code is designed to either start working on a task immediately, or quickly answer with a busy reply - hh_client expects to wait until we are available. Implementing this wait in IDE in non-blocking way requires non-blocking way to wait for the shared memory mutex (and UNIX doesn't make it easy). SEARCH circumvents it by not needing shared memory, but other commands do.
- code paths expecting genv (which is unavailable in IDE process) and ServerEnv.env (which is not the same as IdeEnv.env, though it's prety close by now)

Reviewed By: int3

Differential Revision: D3055495

fb-gh-sync-id: 7f6e33f29ceae6c9ae3754da3a515052e60a7f5c
shipit-source-id: 7f6e33f29ceae6c9ae3754da3a515052e60a7f5c