Support suspend/resume in event-hook
Summary:A lot of times in profiling we want function enters and exits and we want to ignore suspends/resume.
The simplest example is a function callcount profiler which just incremends every time the function is called.

This change does not break the interface, it only adds a flag that when used the "suspend"/"resume" modes are provided.

Broken sandcastle
@override-unit-failures

Reviewed By: jano, markw65

Differential Revision: D3059706

fb-gh-sync-id: ded13ee363962a992446cee83631dc675cb1efb4
shipit-source-id: ded13ee363962a992446cee83631dc675cb1efb4