Backtrace hash should be async aware too
Summary:The goal of the backtrace hash to make backtrace collection and aggregation more performant.
Since now backtraces are async-aware and are very different than a simple PHP stack, it's just plain wrong to use the hash for anything.
This diff moves it near the createBacktrace function and uses the same stack traversal function.

@override-unit-failures
MemoryManager is the only failure and it got fixed

Reviewed By: jano

Differential Revision: D3062282

fb-gh-sync-id: 1b2e2bb73cadf992248bb52191cbd7547ff4d101
shipit-source-id: 1b2e2bb73cadf992248bb52191cbd7547ff4d101