Call hh\idx rather than re-enter the vm
Summary:Right now, the emitter translates calls to hh\idx into the Idx
bytecode, then irgen specializes various cases, and the rest fall
through to a c++ helper called genericIdx.

genericIdx then re-enters the vm to call hh\idx - which is much slower
than just calling hh\idx directly would have been.

I looked at changing the emitter to leave hh\idx as a call, and then
pattern match on that, but there's no guarantee that the FPushFuncD
and the FCall will be in the same tracelet, and its just generally
hard to match on.

So as an alternative, I tried generating the same ir that an FPushD,
FPassC*, FCall would have generated, and it seems to mostly work.

I'm probably missing a bunch of edge cases, and I had to remove an
assert (which will need fixing with something more generic if we go
this route) but looking for feedback on whether this is a good idea.

We're not going to get much of a win from /this/ diff because of my
recent changes to emitIdx, but there are other things that could
benefit from similar bytecode->call transformations (such as casting
objects to strings).

Reviewed By: ottoni

Differential Revision: D3018711

fb-gh-sync-id: 292f6ffc974ec2e8002a6ac19a0e6bc7733a04e5
shipit-source-id: 292f6ffc974ec2e8002a6ac19a0e6bc7733a04e5