Brotli support
Summary:Support for brotli compression.
Until now we have only supported gzip encoding, and now we are adding brotli as
well.
Since only gzip was supported until now, we often made binary decision - we
either do have compression or not. Now we need to consider which type of
compression is enabled / disabled so this code got a bit more involved.
On top of that, due to what seems to be a bug in FF, brotli by default is disabled for
chunked responses. However, there is an option to enable it at runtime by
calling:

  ini_set("brotli.chunked_compression", "on");

Reviewed By: markw65

Differential Revision: D3010183

fb-gh-sync-id: 9048064017fd8d2b6064f236f72e73dda8495413
shipit-source-id: 9048064017fd8d2b6064f236f72e73dda8495413