Begin reordering set / map functions to be more consistent with Core
Summary:I find it ugly and confusing that the Core library functions always take their
collection argument first but the OCaml stdlib doesn't (usually places them in
the middle or as the last argument... it's not even consistent.)

Even though using the Core_map / Core_set libraries would be the ideal
solution, getting them to build under our build system is a pain. In the
meantime, however, I'd like to create a simple layer over the stdlib that
orders arguments a little more sanely, so I don't have to keep looking up
argument orderings online. Also, if and when we do make the Core_map / Core_set
ports, it will be much easier to integrate.

This diff only does it for the SMap / SSets used in `server/`. I will do the
remainder incrementally.

Reviewed By: dlreeves

Differential Revision: D3057331

fb-gh-sync-id: 5659d5aa78062d4b51ec858f6c25174d057a87ce
shipit-source-id: 5659d5aa78062d4b51ec858f6c25174d057a87ce