Fix UNSAFE_EXPR line mismatches
Summary:Multiline `UNSAFE_EXPR` comments currently break line numbers in the rest of the file. Luckily this is exceedingly rare in our code base, so it hasn't been a problem, but it's easy to reproduce:

```
$ cat -n unsafeexpr5.php
     1  <?hh // strict
     2  // Copyright 2004-present Facebook. All Rights Reserved.
     3
     4  function f(string $x): void {
     5    /* UNSAFE_EXPR
     6     * Multiline comments should not mess up the parser.
     7     */
     8    $y = $x + 1;
     9  }
    10
    11  function g(string $x): void {
    12    $y = $x + 1;
    13  }
$ hh
unsafeexpr5.php:16:8,9: Typing error (Typing[4110])
  unsafeexpr5.php:16:8,9: This is a num (int/float) because this is used in an arithmetic operation
  unsafeexpr5.php:15:12,17: It is incompatible with a string
```

This is due to the fact that `L.back` does not restore line numbers.

Fix: instead of parsing `Tunsafeexpr` as a single token, we return only the initial comment-opening token and rely on the parser to consume the rest of the comment.

Reviewed By: int3

Differential Revision: D3070648

fb-gh-sync-id: d5e1e7c0bcfc7fff976e9e7e686f3d70ba2dccf4
shipit-source-id: d5e1e7c0bcfc7fff976e9e7e686f3d70ba2dccf4