Solve optimization issue on PPC64 by improving robustness
Summary:It's not a good idea to simply shut down optimizations if your code is
breaking because the compiler might change in the future. It's wiser
to make your code more robust in order to avoid falling in such cases
that optimization could break your code.

The fixup issue was already solved on upstream
(fa13910bd1f391f6110b1c07cd5a55d265e1b6be) but as discussed above, I
changed it to avoid the optimization issue by using a parent frame
(which tail call optimization doesn't eliminate). There is no drawback
as the called fixupWork will look for a VM frame, which is further back.

Thanks to tuliom for discussing these decisions.
Closes https://github.com/facebook/hhvm/pull/6876

Reviewed By: mxw

Differential Revision: D3034563

fb-gh-sync-id: 732e5db7c41277dfd2d17f1f3075502c5ab9e6a1
shipit-source-id: 732e5db7c41277dfd2d17f1f3075502c5ab9e6a1