Reduce destruction and sweep overhead for HNI classes
Summary:I converted a systemlib class from pure php to HNI, and the cost of
creating the class doubled.

About 20% of the increase was because objects of the class were being
put on a sweep list, even though the class had no sweep method, and no
(explicit) destructor.

I changed it so that we don't register a destructor function if the
class is_trivially_destructible, and we don't make it sweepable if
there's no sweep, and its is_trivially_destructible. I also added
compile time check to see if there is a

  static const bool sweep = false;

declared in the class, and disable sweep in that case too. There's
already a mechanism to not register a sweep method, but thats on the
registerNativeDataInfo call, which can be a long way from the class
definition, so people could easily change the class without realizing
they need to update the call. I've not used it in this diff, but will
in a follow up.

I also audited the registerNativeDataInfo calls and fixed a few
classes to avoid destructor/sweep calls.

Reviewed By: paulbiss

Differential Revision: D3071743

fb-gh-sync-id: 6af7e0b75e1439b9483d7d8d5e0f3284b4282f0c
shipit-source-id: 6af7e0b75e1439b9483d7d8d5e0f3284b4282f0c