Delete RelOffsetData
Summary:According to the comments, we need the BCSPOffset because it's hard to
derive it from the IRSPOffset after irgen is done.  It seems actually
trivial to do so, since we have both the BCSP's offset (in the BCMarker)
and the IRSP's offset (via the DefSP instruction).

Getting rid of RelOffsetData lets us avoid special casing some code for
different stack instructions which would otherwise be the same.

Reviewed By: swtaarrs, paulbiss

Differential Revision: D3076920

fb-gh-sync-id: 0505bd8486bbf456c2fdabf093eaf1c01b3b5829
shipit-source-id: 0505bd8486bbf456c2fdabf093eaf1c01b3b5829