Don't "server busy" the error list requests
Summary:We want the "Server Busy" status for time-sensitive commands, but getting diagnostics is not one of them. The longer term plan is to switch diagnostics to a push-subscribe model, but for now I'm just switching the behavior to mimic what we have now.

In general, this "wait for answers as long as necessary (but don't block the main processing thread)" ability is something that we would need to be able to do for all commands if we want to get to the point of "IDE process answers all hh_client requests too". Then each command would have two variants - one that waits, and one that returns immediately if we can't compute it. We could then perhaps merge IdeJson/IdeServerCall with ServerRpc.

Reviewed By: int3

Differential Revision: D3073041

fb-gh-sync-id: 41616459e91e3128a220d32c1da45be38ab8cc7c
shipit-source-id: 41616459e91e3128a220d32c1da45be38ab8cc7c