Hack: make the HHI tarball generation umask-agnostic
This is a follow-up to D2834659, which aimed to make the HHI tarball
generation deterministic/reproducible. That commit unfortunately didn't
catch the case of a varying umask, which results in the tarball having
different modes for files/directories depending on the builder's umask.
Pass --mode to tar, to force the mode to a static one.