Delete jit::Location
Summary:Nobody cares about the non-{Stack,Local} uses of jit::Location
anymore, in which case it's functionally equivalent to
RegionDesc::Location.  Let's delete the former, and rename the
latter to it in a later diff.

(We did check to see if the predictedType() for, e.g., literals or This
were known DataTypes, but they /always/ were, so checking them is
spurious.)

Reviewed By: swtaarrs

Differential Revision: D3085227

fb-gh-sync-id: be9411c8309304ffc1f774b50f5c0c5847d05b5b
shipit-source-id: be9411c8309304ffc1f774b50f5c0c5847d05b5b