Refactor scheduler [1/?]
Summary:I'm looking at the current IDE scheduler code to try and get a sense of how difficult it would be to convert it to run all the operations in the typechecker process as well. As part of that I'm trying to see how the existing code can be better factored.

This diff removes IdeIdle in favor of having services call the scheduler directly. IdeIdle (like ServerIdle) is like a poor-man's scheduler, but now that we have a better one the additional layer of indirection seems unnecessary. It also moves around a bunch of modules in order to avoid circular dependencies.

Reviewed By: dabek

Differential Revision: D3087042

fb-gh-sync-id: fee4b780444e55ff479cdde91242151b13beb46d
shipit-source-id: fee4b780444e55ff479cdde91242151b13beb46d