Don't spawn the IDE process for now
Summary:Per conversations with Jez/Julien/Peter we're going to do the Nuclide work more gradually (see summary of D3095598), and possibly merge the IDE process (along with it's scheduler) into the main process.

IDE process was supposed to be simple - run on the side and just read shared memory. Actual implementation revealed that it's not enough to answer the queries, and amount of dependencies between typechecker and IDE made the whole thing much more complicated that I wanted and what is manageable. It should also unblock lazy-decl launch.

For now just commenting the key parts that spawn the IDE process and reverting {D3055495}. Leaving the scheduler/JSON protocol code in place to hopefully eventually be merged into main process.

Reviewed By: int3

Differential Revision: D3095831

fb-gh-sync-id: f9278ad8bac8732b8f2211b37eea5372c8f75d63
shipit-source-id: f9278ad8bac8732b8f2211b37eea5372c8f75d63