Don't touch CodeCache::data() until final code emission
Summary:This was the only remaining problem (that I'm aware of) in the way of
running most of the vasm pipeline without the write lease. Previously, we were
bump-allocating space in globalData in code-gen-x64.cpp. This has been replaced
with a mechanism to temporarily allocate the relevant data on the heap,
relocating it to globalData right before code emission. See comments inline for
a full explanation of what's going on.

I also fixed a non-atomic global counter in vasm-xls.cpp.

Reviewed By: mxw

Differential Revision: D3027581

fb-gh-sync-id: c462c5e307648eecc2b7e5dbd6298751ef53cbae
fbshipit-source-id: c462c5e307648eecc2b7e5dbd6298751ef53cbae