Cleanup TransLocMaker::rec()
Summary:This method had some logic to handle the cases where cold and frozen are
aliases, which had the purpose of avoiding printing that code twice in tc-print
in such cases.  However, this logic was broken recently.  But instead of fixing
it here, I now added checks for those conditions in tc-print.  So just get rid
of the silly logic remaining here.

Reviewed By: alexmalyshev

Differential Revision: D3101524

fb-gh-sync-id: a27cab58682607066acdb972344da25ce04e6e0c
fbshipit-source-id: a27cab58682607066acdb972344da25ce04e6e0c