Rename asio STATE_SCHEDULED to STATE_READY
Summary:Once a waithandle is unblocked, it is ready to execute.
At any given time, a READY waithandle may exist in more
than one context's queue, or in no context's queue at all.
Since it's confusing to think of something "scheduled"
but not in any queue, READY is a more precise state name.

Reviewed By: jano

Differential Revision: D3105695

fb-gh-sync-id: 71d45f9e646978082dcf7183506c9a43f4ca8571
fbshipit-source-id: 71d45f9e646978082dcf7183506c9a43f4ca8571