Refactor SlotState update routines
Summary:Minstr guards necessitate either copying or reusing lots of this code.
Currently we have two copies, and I really didn't want to add a third.

This diff also fixes a bug in our attempts to preserve predicted types
when we StLoc (or similar) the result of a LdLoc.  We were assuming
before that when setting a local's value in FrameStateMgr::update(),
that if the src was a LdLoc, it would always be from the same frame.
This is not true in the presence of inlining, when the LdLoc could
easily be relative to our parent's frame.  This was causing an out-of-
bounds vector access, and probably incorrect predictions; the fix is
simply to look in the FrameState stack for the proper frame locals.

Reviewed By: swtaarrs

Differential Revision: D3086652

fb-gh-sync-id: c0874038d967fca2b0e60b4f54e3164283025eaa
fbshipit-source-id: c0874038d967fca2b0e60b4f54e3164283025eaa