Use a recoverable fatal on curl_init_pooled
Summary:curl_init_pooled has an unfortunate failure mode, when the pool is
exhausted the entire request fatals. This prevents the application from
gracefully handling the failure. The application could disable the
feature rather than canceling the entire request. In general fatal
errors should be reserved for cases where it is impossible to continue
processing the request, not simply because something went wrong. This
switches the fatal to a RuntimeException.
Closes https://github.com/facebook/hhvm/pull/6920

Reviewed By: Orvid

Differential Revision: D3066969

fb-gh-sync-id: 7c3b92cad44627cb3ba67c88f847e5e500522ab0
fbshipit-source-id: 7c3b92cad44627cb3ba67c88f847e5e500522ab0