Refactor AsyncFunction/GeneratorWaitHandle::resume()
Summary:Waithandles in an AsioContext's queue can change from READY
to antoher state while they're waiting in the queue. resume()
handles the situation by stopping early; in that case we're
destorying the WH pointer owned by the queue, so we decref.

Refactor the code to do the state check and decref in AsioContext
(close to where we're destroying the pointer).

Differential Revision: D3116121

fb-gh-sync-id: 9bb8e463c89a16f73cbd46e9ee3064cc9a349ce6
fbshipit-source-id: 9bb8e463c89a16f73cbd46e9ee3064cc9a349ce6