Shared memory can be allocated at any address
Summary:This was how our shared memory used to be allocated.  When I made the original
saved state -- that contained the hashtable -- I made the address fixed so that
I wouldn't have to update the pointers in the hashtable when loading. However,
mini saved state doesn't store the hashtable any more, so this is now
unnecessary.

Reviewed By: dlreeves

Differential Revision: D3121564

fb-gh-sync-id: ab01d7b72e32e65f147dbf148e9542e12835aa17
fbshipit-source-id: ab01d7b72e32e65f147dbf148e9542e12835aa17