Remember to detach the XDebugHook on request remote mode
Summary:In the process of initializing xdebug on a request, we first attach the
XDebugHook, which then tries to set up the XDebugServer.  One of the things the
XDebugHook does is set the Debugger surprise flag, which is sticky.  If the
XDebugServer fails to connect to a client, it gets discarded, but the XDebugHook
sticks around as does the surprise flag.

This is actually intended behavior, but only when the xdebug remote mode is set
to "jit", which means a debugger can attach whenever a request throws an
exception.  However we're not using that mode, we're using "req", which just
means try to connect to a debugger at the start of a request.

All that's needed is to detach the XDebugHook when the XDebugServer fails to
initialize and we're running under "req".  That way we won't fail every surprise
check ever.

Reviewed By: paulbiss

Differential Revision: D3122683

fb-gh-sync-id: 77bf3485de95b3b6bba66c1d32696c41b7ffe323
fbshipit-source-id: 77bf3485de95b3b6bba66c1d32696c41b7ffe323