Add test for incremental handling of files that failed decl
Summary:When a file generates errors in the decl phase, we make the very
conservative assumption that we know nothing about its dependencies.
So we redeclare it every time a file changes, even if the changed file
is wholly unrelated to the contents of the erroneous file.

I can't actually think of any example where this is necessary, but
until we have a good argument for why it isn't, it's probably
reasonable to ensure this behavior is not accidentally broken.

The current implementation of lazy decl breaks it; I plan to fix that
in a follow-up commit.

Reviewed By: dlreeves

Differential Revision: D3114537

fb-gh-sync-id: d79901b491f85d6545424b7e6c96d55fc11442ce
fbshipit-source-id: d79901b491f85d6545424b7e6c96d55fc11442ce