Store lazy decl errors in shared memory
Summary:Lazy decl makes it more complicated to keep track of all the errors that
declaration could produce. Since any command can now result in the declaration
code running, we must be able to record the errors in some global state. This
diff adds some shared memory lists for that purpose.

Whenever decl gets invoked lazily in any of the workers, it records its errors
in those lists. At the end of every recheck loop, the master retrieves the
contents of those lists (and empties them).

Strictly speaking, this may not be entirely necessary. If we are incrementally
updating from a saved state, all the files that we are not eagerly declaring
and rechecking should in theory be error-free. However, there is no static
guarantee of this, and we would need some way to record these errors at
runtime. The plumbing required to do this logging requires adding global state
too, and doesn't seem any simpler than this approach.

Looking further ahead: at this point, we have two ways that errors get reported
to the master -- via the return value of the workers and via the shared lists.
It would be nice to unify them.

Reviewed By: dlreeves

Differential Revision: D3118402

fb-gh-sync-id: 559526a0d6113a6aa6e168eec269b507e9cf6ec0
fbshipit-source-id: 559526a0d6113a6aa6e168eec269b507e9cf6ec0