Double the number of keys allowed in the shared hashtable
Summary:Increase HASHTBL_POW from 25 to 26 when OSS_SMALL_HH_TABLE_POWS is NOT defined. This increases the memory used by the shared hash table by 32 MB, which does not seem to make much of a discernible difference on large machines.

The motivation for this is to allow static analysis plug-ins to make more aggressive use of the shared hash table.

@override-unit-failures

Reviewed By: int3

Differential Revision: D3135422

fb-gh-sync-id: 624e6dd5498ca613c23476e7a2341fe4fa2804bc
fbshipit-source-id: 624e6dd5498ca613c23476e7a2341fe4fa2804bc