Add a filter_map option to Trie search
Summary: Autocomplete needs to be fast, and to achieve this it short-circuits computations after getting first 100 results. I want to replace current "go through all class/function names looking for a match" with a trie lookup, but in order to be able to short-circuit it the same way we need to generate the results as we go, instead of getting the bach of results first and then mapping over them.

Reviewed By: int3

Differential Revision: D3116040

fb-gh-sync-id: ee7c6d17bfba26e2bec8d5cd1876512e8d0c6a5f
fbshipit-source-id: ee7c6d17bfba26e2bec8d5cd1876512e8d0c6a5f