Add tool to generate type-scanners automatically
Summary:Add a tool which examines the debug information from an executable and
uses it to automatically generate type-scanners. Add the build steps to
automatically run this during the HHVM build, compile the scanners into a shared
object, and embed that shared object into the executable itself.

Since the tool currently takes almost a minute to run, unless the
HHVM_BUILD_TYPE_SCANNERS environment variable is set, no analysis will be done,
and the generated type-scanners will be all conservative.

override-unit-failure

Reviewed By: edwinsmith

Differential Revision: D3025598

fb-gh-sync-id: 7e350a6a67e3b555c13f7c835828a4ae6a8f1a5d
fbshipit-source-id: 7e350a6a67e3b555c13f7c835828a4ae6a8f1a5d