Expose isValid and add catchable exceptions
Summary:We saw `mysql_connection_ != nullptr` checks firing in PHP and we have no proactive way to verify the underlying connection.
This exposes isValid as a way to check the data inside `AsyncMysqlConnection`. When a query times out, the Squangle layer closes that connection because it's no longer safe to access it, the extension layer doesn't do anything special thus allowing the above layer to try to access, this causes an exception that's not catchable in PHP.

This diff only adds LOGs and verification where it's safe to. Another diff will make the extension treat the timeout case.

Reviewed By: anca-agape

Differential Revision: D3158489

fb-gh-sync-id: 4e729da99bed9da541e241c4de018092d1259027
fbshipit-source-id: 4e729da99bed9da541e241c4de018092d1259027