Add a script to annotate printir output
Summary:Given the output of TRACE=printir, and the output of perf script from
the same run, this will annotate the ir dump with the number of events
per address (much like "annotate" within perf report does for C++ functions).

Example usage:

  TRACE=printir perf record -e instructions -- hhvm myscript.php
  perf script > myscript.perf
  hphp/tools/perf-annot --event=instructions /tmp/hphp.log myscript.perf > hphp.log.annotated

It can similarly be used in server mode, with suitable modifications to the above sequence.

Reviewed By: ottoni

Differential Revision: D3156779

fb-gh-sync-id: 42ee76829c70ff7502a711c21a1fab91c8d81307
fbshipit-source-id: 42ee76829c70ff7502a711c21a1fab91c8d81307