Unit tests for IdentifySymbolService
Summary:I will be changing it, and it's not very well tested code.

We don't have a good way to unit test methods that take a file position as argument - autocomplete uses magical AUTO332 marker for that, but this is not very convenient. To solve this I'm adding a .flags file - similarly to .no_format they are per-test files with additional flags passed to the test binary, just like per-dir HH_FLAGS file does. This addition breaks the invariant (?) that entire unit test input is contained in a single file, but it seems that the alternative to it is to not have tests at all, so...

Reviewed By: int3

Differential Revision: D3164938

fb-gh-sync-id: 643992b8dca5f3d42cf7a27431e5493f4a17cbb0
fbshipit-source-id: 643992b8dca5f3d42cf7a27431e5493f4a17cbb0