Hook up type-scanners to GC
Summary:Add the optional ability to use the type-scanners during garbage
collection. This is controlled by the new runtime option
Eval.EnableGCTypeScan=1. If this option is used without a
HHVM_BUILD_TYPE_SCANNERS, all conservative scanning will be done. The existing
scanners and the type-scan mechanism will continue to live side-by-side for a
while, as right now the type-scanners are only supported for resources and
malloc allocations.

Reviewed By: edwinsmith

Differential Revision: D3144690

fb-gh-sync-id: bba52af5237ed147ebe643647eb0cbf964d6e755
fbshipit-source-id: bba52af5237ed147ebe643647eb0cbf964d6e755